Fixes Issue #2 removed empty python file, moves all changes to main.py and deletes python2.py and python3.py, updates requirements to play video and load images
