+++ main.py
@@ -5,6 +5,8 @@ from kivymd.uix.button import MDRaisedButton, MDFlatButton
 from kivy.lang import Builder
 from kivymd.app import MDApp
 from kivy.uix.gridlayout import GridLayout
+from kivy.uix.videoplayer import VideoPlayer
+from kivy.uix.scrollview import ScrollView
 
 # Used for most MD apps
 from kivymd.app import MDApp
@@ -19,26 +21,36 @@ from kivy.core.window import Window
 # Example Android screen size
 Window.size = (2224, 1668)
 
+# Set path for video file
+import os
+video_file_path = os.path.join(os.path.dirname(__file__), 'shrekrizz.mp4')
+
 # Class for each screen and its canvas
 class HomeScreen(Screen):
-    pass
+
     class HomeScreenCanvas(FloatLayout):
         pass
 
 class QAScreen(Screen):
-    pass
+
     def submit_question(self):
             question = self.ids.question_input.text
             if question:
                 self.ids.qa_container.add_widget(MDLabel(text=question))
                 self.ids.question_input.text = ''
+
     class QAScreenCanvas(FloatLayout):
         pass
 
 class QuizScreen(Screen):
-    pass
+
     class QuizScreenCanvas(FloatLayout):
-        pass
+        def on_raindrop_select(self, instance):
+            instance.md_bg_color = [0, 1, 0, 1]  # Glowing border effect
+
+        def submit_answer(self):
+            # Logic to go to next question and update interface
+            pass
 
 class LessonScreen(Screen):
     pass
@@ -92,8 +104,66 @@ class WritingScreen(Screen):
     pass
     class WritingScreenCanvas(FloatLayout):
         pass
+
+class LoginScreen(Screen):
+    pass
+    def __init__(self, **kwargs):
+        super().__init__(**kwargs)
+
+        layout = BoxLayout(orientation='vertical', spacing=10, padding=20)
+
+        self.username_input = TextInput(
+            hint_text="Username",
+            multiline=False,
+            size_hint=(0.6, 0),  # Adjust width and height
+            pos_hint={"center_x": 0.5, "center_y": 0.5}  # Center horizontally
+        )
+
+        self.password_input = TextInput(
+            hint_text="Password",
+            password=True,
+            size_hint=(0.6, 0),  # Adjust width and height
+            pos_hint={"center_x": 0.5, "center_y": 0.5}  # Center horizontally
+        )
+
+        login_button = MDRaisedButton(text="Login", on_press=self.validate_login)
+
+        layout.add_widget(self.username_input)
+        layout.add_widget(self.password_input)
+        layout.add_widget(login_button)
+
+        self.add_widget(layout)
+
+    def validate_login(self, instance):
+        username = self.username_input.text
+        password = self.password_input.text
+        if username == "admin" and password == "1234":  # Example credentials
+            print("Login successful!")  # You can navigate to another screen here
+        else:
+            print("Invalid credentials")
+    class LoginScreenCanvas(FloatLayout):
+        pass
+
+
+class LessonScreenone(Screen):
+
+    def __init__(self, **kwargs):
+        super().__init__(**kwargs)
+        layout = BoxLayout(orientation="vertical")
+        video = VideoPlayer(source=video_file_path, state="play", options={"allow_stretch": True}, size_hint=(1, 1))
+        layout.add_widget(video)
+        self.add_widget(layout)
+
+    class LessonScreenoneCanvas(FloatLayout):
+        pass
+
+class Lessonscreenoverview(Screen):
+    pass
+    class LessonscreenoverviewCanvas(FloatLayout):
+        pass
+
     
-class AboutMeApp(MDApp):
+class IntercalmSpeechApp(MDApp):
     def build(self):
         # Light Mode
         # self.theme_cls.primary_palette = "Blue"
@@ -112,10 +182,13 @@ class AboutMeApp(MDApp):
         self.sm.add_widget(LessonScreen(name='lessonscreen'))
         self.sm.add_widget(FlashcardScreen(name='flashcardscreen'))
         self.sm.add_widget(WritingScreen(name='writingscreen'))
+        self.sm.add_widget(LessonScreenone(name='lessonscreenone'))
+        self.sm.add_widget(LoginScreen(name='login'))
+        self.sm.add_widget(Lessonscreenoverview(name='lessonscreenoverview'))
         return self.sm
 
     def go_home(self):
         self.sm.current = 'home'
 
 if __name__ == '__main__':
-    AboutMeApp().run()
+    IntercalmSpeechApp().run()
