+++ requirements.txt
@@ -1,2 +1,4 @@
 kivy
 kivymd
+ffpyplayer
+pillow
